Move this to utilities.
