+++ src/SparkPost/ListSubaccountResponse.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using SparkPost.Utilities;
 
 namespace SparkPost
 {
@@ -12,7 +13,7 @@ namespace SparkPost
             var result = new ListSubaccountResponse();
             LeftRight.SetValuesToMatch(result, response);
 
-            var results = JsonStuff.DeserializeObject<dynamic>(response.Content).results;
+            var results = Jsonification.DeserializeObject<dynamic>(response.Content).results;
             var subaccounts = new List<Subaccount>();
             foreach (var r in results)
                 subaccounts.Add(ConvertToSubaccount(r));
