+++ src/SparkPost/ListWebhookResponse.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using SparkPost.Utilities;
 
 namespace SparkPost
 {
@@ -11,7 +12,7 @@ namespace SparkPost
             var result = new ListWebhookResponse();
             LeftRight.SetValuesToMatch(result, response);
 
-            var results = JsonStuff.DeserializeObject<dynamic>(result.Content).results;
+            var results = Jsonification.DeserializeObject<dynamic>(result.Content).results;
             var webhooks = new List<Webhook>();
             foreach(var r in results)
                 webhooks.Add(ConvertToAWebhook(r));
