+++ src/SparkPost/RequestMethods/PutAndPostAreTheSame.cs
@@ -1,6 +1,8 @@
 ﻿using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
+using SparkPost.Utilities;
 
 namespace SparkPost.RequestMethods
 {
@@ -27,8 +29,7 @@ namespace SparkPost.RequestMethods
 
         private static string SerializeObject(object data)
         {
-            return JsonStuff.SerializeObject(data,
-                new JsonSerializerSettings {TypeNameHandling = TypeNameHandling.None});
+            return Jsonification.SerializeObject(data);
         }
     }
 }
