+++ src/SparkPost/Suppressions.cs
@@ -4,6 +4,7 @@ using System.Net;
 using System.Threading.Tasks;
 using System.Web;
 using SparkPost.RequestSenders;
+using SparkPost.Utilities;
 
 namespace SparkPost
 {
@@ -38,7 +39,7 @@ namespace SparkPost
             var response = await requestSender.Send(request);
             if (response.StatusCode != HttpStatusCode.OK) throw new ResponseException(response);
 
-            var results = JsonStuff.DeserializeObject<dynamic>(response.Content).results;
+            var results = Jsonification.DeserializeObject<dynamic>(response.Content).results;
 
             return new ListSuppressionResponse
             {
@@ -63,7 +64,7 @@ namespace SparkPost
                 throw new ResponseException(response);
 
             dynamic results = response.StatusCode == HttpStatusCode.OK
-                ? JsonStuff.DeserializeObject<dynamic>(response.Content).results
+                ? Jsonification.DeserializeObject<dynamic>(response.Content).results
                 : null;
 
             return new ListSuppressionResponse
