+++ src/SparkPost/Transmissions.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using SparkPost.RequestSenders;
+using SparkPost.Utilities;
 
 namespace SparkPost
 {
@@ -31,7 +32,7 @@ namespace SparkPost
             var response = await requestSender.Send(request);
             if (response.StatusCode != HttpStatusCode.OK) throw new ResponseException(response);
 
-            var results = JsonStuff.DeserializeObject<dynamic>(response.Content).results;
+            var results = Jsonification.DeserializeObject<dynamic>(response.Content).results;
             return new SendTransmissionResponse()
             {
                 Id = results.id,
@@ -65,7 +66,7 @@ namespace SparkPost
 
             try
             {
-                var results = JsonStuff.DeserializeObject<dynamic>(response.Content).results;
+                var results = Jsonification.DeserializeObject<dynamic>(response.Content).results;
                 if (results.transmission == null) return transmissionResponse;
 
                 transmissionResponse.Id = results.transmission.id;
